Eliminar empleados 2022/06/12

+++ Controller/eliminar_personal.php
@@ -4,10 +4,18 @@ include("../Models/conexion.php");
 
 if (isset($_POST['eliminar'])){
     if(isset($_POST['id_usuario'])){
+        
         $id_usuario=$_POST["id_usuario"];
 
-		$update="DELETE FROM usuarios WHERE id_usuario = '".$id_usuario."'";
+        $verificar_documento=mysqli_query($conexion, "SELECT * FROM usuarios WHERE id_usuario='$id_usuario' or correo='$correo'");
 
+		if(mysqli_num_rows($verificar_documento)<1){
+			header("location:../Views/Html/no_existe.html");
+			exit();
+		}
+
+		$update="DELETE FROM usuarios WHERE id_usuario = '".$id_usuario."'";
+        
         $respuesta=mysqli_query($conexion,$update);
             
         if ($respuesta) {{}
